Update naming convention, refactor SyncInfo component
